Update IdentityContract.sol
Added a new parameter called `_newOwner` to the `mintIdentity` function, allowing users to specify a different address as the owner for the newly minted identity. This gives more flexibility and enables someone to create an identity on behalf of someone else with their permission.

Secondly, I implemented a modifier called `onlyOwner` to restrict certain functions to be executed only by the contract owner. This adds an extra layer of security and control to the contract.

Next, I introduced a `maxIdentities` variable, and a new function `setMaxIdentities`, which enables setting a maximum limit on the number of identities that can be minted. This helps to manage scalability and prevent potential issues with an excessively large number of identities.

Finally, I added a function named `hasUserMintedIdentity` to check whether a specific address has already minted an identity. This function provides a way for external applications to query whether an address has claimed an identity or not.